+++ backend/package.json
@@ -12,7 +12,7 @@
     "nodemon": "^3.1.9"
   },
 "scripts": {
-  "start": "node server.js",
-  "dev": "nodemon server.js"
+  "start": "node index.js",
+  "dev": "nodemon index.js"
 }
 }
